refactor: Prepare for splitting the application DI
Prepare for splitting the application construction into a loading and construction phase by renaming
the existing factory and model to the target names.
